requirements.txt
Flask
Flask-PyMongo
Flask-JWT-Extended
gunicorn
python-dotenv